mlx4_core: Extend num_mtt in dev caps to avoid overflow.
Some legitimate combinations of log_num_mtt and log_mtts_per_seg
cause overflow in the calculation of the num_mtt when initializing
the HCA which causes Kernel panic. Changed the variable to be 'u64'
instead of 'int' to avoid the overflow and made the needed changes
to support the new type.

(Ported from Mellanox OFED 2.4)


Orabug: 27510485

(cherry picked from commit a15f6fc5eb000c0287afdb03c2e696b73b634012)
cherry-pick-repo=linux-uek.git

+++ include/linux/mlx4/device.h
@@ -577,7 +577,7 @@ struct mlx4_caps {
 	int			num_comp_vectors;
 	int			num_mpts;
 	int			max_fmr_maps;
-	int			num_mtts;
+	u64			num_mtts;
 	int			fmr_reserved_mtts;
 	int			reserved_mtts;
 	int			reserved_mrws;
